optimize ut
